[#147] Propose a specialized trait for random access iterators.
The function contiguos_block_list returns a sequence of triples (l,b,e) where l
is the Locality, b is a global iterator to the start of a local block in l, and
e i the end of a local block in l.  The list contains one element for each and
every blocks that are in the input range.